[fix] Added comparisons redered correctly
Comparisons are added to the chart conform to the main Stock/Security selected. Where dates do not exist
a blank value is rendered and the corresponding dates are now added for the comparison that avoid being shifted
left/right due to main stock max/min dates starting before or after the comparison.